feat: Add IpAsnMap::new() & implement Default
Implements `Default` for `IpAsnMap` and adds a convenience method
`IpAsnMap::new()` to allow for the simple creation of an empty map.

This improves ergonomics by providing a more intuitive and direct way to
instantiate an empty map, which is useful for scenarios where a map
might be populated conditionally or used as a placeholder.